+++ packages/cp-listing-service/src/services/events.service.ts
@@ -25,16 +25,26 @@ export const serviceEvents = async (channel: Channel) => {
   }
 };
 
-export const sendToServiceQueues = async (channel: Channel, message: any, services: string[] = []) => {
+export const sendToServiceQueues = async (
+  channel: Channel,
+  message: any,
+  services: string[] = []
+) => {
   const promises: any[] = [];
   services.forEach((service) => {
-    promises.push(channel.sendToQueue(service, Buffer.from(JSON.stringify(message)), { persistent: true }));
+    promises.push(
+      channel.sendToQueue(service, Buffer.from(JSON.stringify(message)), { persistent: true })
+    );
   });
   const results = await Promise.all(promises);
   console.log(results);
 };
 
-export const getServiceQueues = async (redis: RedisConnection, scope: string, services: string[] = []) => {
+export const getServiceQueues = async (
+  redis: RedisConnection,
+  scope: string,
+  services: string[] = []
+) => {
   await redis.client.connect();
   const registeredQueues = await redis.client.sMembers(`${scope}-queues`);
   let relevantQueues: string[] = [];